Update README.md
With over 3 years of experience in WordPress development, I specialize in creating visually stunning and highly functional websites that deliver results. My journey in web development is rooted in a passion for blending design with cutting-edge technology, ensuring each project I take on not only looks great but also performs seamlessly.

I hold a BS in Bioinformatics, which has equipped me with a unique perspective on problem-solving and an analytical approach to design and development. Whether you're looking to build a sleek business website, an engaging landing page, or a full-fledged e-commerce platform, I'm here to turn your vision into reality.